fix: length should not be trunc due to int size being 32 on 32bit

+++ r_strlen.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int r_strlen(char *a)
+size_t r_strlen(char *a)
 {
     size_t l = 0;
     while (*a++ != '\0')
